modularization
TODO:
- report writing
- readme update
