+++ training/train_elevator_rl.py
@@ -1,6 +1,6 @@
 # train_elevator_rl.py
-from elevator_env import ElevatorEnv
-from elevator_dqn import ElevatorDQN, ElevatorDDQN, ElevatorTDQN
+from agent.elevator_env import ElevatorEnv
+from agent.elevator_dqn import ElevatorDQN, ElevatorDDQN, ElevatorTDQN
 import matplotlib.pyplot as plt
 import numpy as np
